+++ spec/active_record/connection_adapters/oracle_enhanced_context_index_spec.rb
@@ -174,6 +174,8 @@ describe "OracleEnhancedAdapter context index" do
   describe "on multiple tables" do
     before(:all) do
       @conn = ActiveRecord::Base.connection
+      @oracle12c = !! @conn.select_value(
+                      "select * from product_component_version where product like 'Oracle%' and to_number(substr(version,1,2)) = 12")
       create_tables
       class ::Post < ActiveRecord::Base
         has_many :comments, dependent: :destroy
@@ -192,10 +194,13 @@ describe "OracleEnhancedAdapter context index" do
     end
 
     after(:each) do
+      @conn.remove_context_index :posts, name: 'post_and_comments_index' rescue nil
+      @conn.remove_context_index :posts, index_column: :all_text rescue nil
       Post.destroy_all
     end
 
     it "should create multiple table index with specified main index column" do
+      pending "It always fails when Oracle 12c 12.1.0 used." if @oracle12c
       @conn.add_context_index :posts,
         [:title, :body,
         # specify aliases always with AS keyword
@@ -210,10 +215,10 @@ describe "OracleEnhancedAdapter context index" do
       ["aaa", "bbb", "ccc", "ddd", "eee", "fff"].each do |word|
         Post.contains(:all_text, word).to_a.should == [@post]
       end
-      @conn.remove_context_index :posts, name: 'post_and_comments_index'
     end
 
     it "should create multiple table index with specified main index column (when subquery has newlines)" do
+      pending "It always fails when Oracle 12c 12.1.0 used." if @oracle12c
       @conn.add_context_index :posts,
         [:title, :body,
          # specify aliases always with AS keyword
@@ -232,7 +237,6 @@ describe "OracleEnhancedAdapter context index" do
       ["aaa", "bbb", "ccc", "ddd", "eee", "fff"].each do |word|
         Post.contains(:all_text, word).to_a.should == [@post]
       end
-      @conn.remove_context_index :posts, name: 'post_and_comments_index'
     end
 
     it "should find by search term within specified field" do
@@ -252,7 +256,6 @@ describe "OracleEnhancedAdapter context index" do
       Post.contains(:all_text, "ccc within comment_body").to_a.should be_empty
       Post.contains(:all_text, "ddd within comment_body").to_a.should == [@post]
       Post.contains(:all_text, "ddd within comment_author").to_a.should be_empty
-      @conn.remove_context_index :posts, index_column: :all_text
     end
 
   end
@@ -378,6 +381,9 @@ describe "OracleEnhancedAdapter context index" do
 
     describe "with table prefix and suffix" do
       before(:all) do
+        @conn = ActiveRecord::Base.connection
+        @oracle12c = !! @conn.select_value(
+                        "select * from product_component_version where product like 'Oracle%' and to_number(substr(version,1,2)) = 12")
         ActiveRecord::Base.table_name_prefix = 'xxx_'
         ActiveRecord::Base.table_name_suffix = '_xxx'
         create_tables
@@ -402,6 +408,7 @@ describe "OracleEnhancedAdapter context index" do
       end
 
       it "should dump definition of multiple table index with options" do
+        pending "It always fails when Oracle 12c 12.1.0 used." if @oracle12c
         options = {
           name: 'xxx_post_and_comments_i',
           index_column: :all_text, index_column_trigger_on: :updated_at,