+++ History.md
@@ -1,3 +1,57 @@
+## 1.5.5 / 2014-05-23
+
+* Enhancements
+ * Oracle NUMBER datatype can be handled as Rails :float datatype [#418]
+   - Default NUMBER datatype handled as :decimal to keep compatibility
+   - Configured by setting `self.number_datatype_coercion = :float` 
+ * Add link to supported Oracle database version, JDK and Oracle JDBC Driver version [#438]
+ * Support `without_prepared_statements?` to handle `unprepared_statement` [#447]
+
+* Bug Fix
+  * Associations with name `record` do not work correctly since Rails 4 [#435]
+  * Skip another Oracle Text test when Oracle 12c used [#437]
+  * Tag bind params with a bind param object [#444]
+
+## 1.5.4 / 2014-03-25
+
+* Enhancements
+ * Support Rails 4.1.0.rc2
+ * Allow Java 8 to run with jruby [#383]
+
+* Bug Fix
+  * Fix db:schema:dump when foreign key column name is not 'id' [#409]
+  * Fix schema dump works when non Oracle adapter used [#428]
+
+## 1.5.3 / 2014-03-04
+
+* Enhancements
+ * Supports Rails 4.1.0.rc1
+ * Support rails/rails#13886 by chainging select_rows arguments [#415]
+
+* Bug Fix
+  * Fix ORA-01008: not all variables bound [#422]
+
+## 1.5.2 / 2014-01-24
+
+* Enhancements
+ * Supports Rails 4.1.0.beta1
+ * Support Rails 4 Database Tasks [#404]
+ * Create sequence when add primary_key column [#406]
+ * Move `SchemaCreation` to its own file [#381]
+ * Remove unused OracleEnhancedColumnDefinition [#382]
+ * Log bind variables after they were type casted [#385]
+ * Remove add_order_by_for_association_limiting! [#388]
+ * Support named savepoints [#389]
+ * Support self.extract_value_from_default [#395]
+ * Remove oracle_enhanced_core_ext.rb [#397]
+ * Remove unused to_sql_with_foreign_keys and lob_columns [#398]
+ * Remove ruby-oci8 v1 code [#405]
+
+* Bug Fix
+  * Move add_column_options! into SchemaCreation class [#384]
+  * Add options_include_default! [#384]
+  * Use OCI8::Metadata::Base#obj_link [#399]
+
 ## 1.5.1 / 2013-11-30
 
 * Enhancements